test: refactor: move `add` function to common.ts
